事务

+++ src/main/java/com/plaso/dao/UserDao.java
@@ -8,5 +8,5 @@ public interface UserDao {
 
     User selectUser(int id);
 
-    Long save(User user);
+    int save(User user);
 }
